Refactor central_columns functions to static methods
